+++ .version
@@ -1,1 +1,1 @@
-BUILD_VERSION=1.0.190
+BUILD_VERSION=1.0.192
